bookmarks: Remove public_node and introduce a fetch module
